tests/wpt/tests/css/css-font-loading/fontfacesetloadevent-constructor.html

<!DOCTYPE html>
<title>FontFaceSetLoadEvent constructor</title>
<link rel="help" href="https://drafts.csswg.org/css-font-loading/#fontfacesetloadevent">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(function() {
    var ff = [];
    var e = new FontFaceSetLoadEvent('type');
    assert_array_equals(e.fontfaces, ff);
    assert_not_equals(e.fontfaces, ff);
  }, 'FontFaceSetLoadEvent constructor without FontFaceSetLoadEventInit dictionary');

  test(function() {
    var ff = [ new FontFace('family', 'src') ];
    var e = new FontFaceSetLoadEvent('type', { fontfaces: ff });
    assert_array_equals(e.fontfaces, ff);
    assert_not_equals(e.fontfaces, ff);
  }, 'FontFaceSetLoadEvent constructor with FontFaceSetLoadEventInit dictionary');
</script>
