LayoutTests/imported/w3c/web-platform-tests/css/css-font-loading/fontfacesetloadevent-constructor-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didReceiveTitle: FontFaceSetLoadEvent constructor
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame

FAIL FontFaceSetLoadEvent constructor without FontFaceSetLoadEventInit dictionary Can't find variable: FontFaceSetLoadEvent
FAIL FontFaceSetLoadEvent constructor with FontFaceSetLoadEventInit dictionary Can't find variable: FontFaceSetLoadEvent
